net/http: on Transport body write error, wait briefly for a response
The http.RoundTripper interface says you get either a *Response, or an
error.

But in the case of a client writing a large request and the server
replying prematurely (e.g. 403 Forbidden) and closing the connection
without reading the request body, what does the client want? The 403
response, or the error that the body couldn't be copied?

This CL make the Transport give an N millisecond advantage to
responses over body write errors.